split env var loading into web/native, move out of constants
